Refactor share related code into smaller chunks
